refactor(server): finish the manage_operations extraction
A manage_operations/ directory has sat beside manage_operations.ts since
the extraction was started, holding 627 lines while the file kept 3,488 —
the highest branch count in the repo. By shape it was indistinguishable
from the finished extractions next to it (manage_connections.ts at 140
lines over 4,229; get_content.ts at 20 over 3,289).

Follows that same template:

  handlers/approvals.ts      1,918  builder + entity-change approval flow
  handlers/execute.ts          806  local, MCP, and OpenAPI execution
  handlers/list-available.ts    742  discovery, readiness, next-action
  handlers/runs.ts             229  run records and activity
  handlers/shared.ts            28  cross-handler helpers
  schemas.ts                    15  action contracts
  manage_operations.ts          65  facade

The facade re-exports all eleven previously exported symbols, so none of
the 20+ importers change. No handler imports the facade back, so the
circular-init hazard that put MANAGE_BEHAVIORS_ACTION_KEY in TDZ is not
reintroduced; that deferral and its comment moved intact into approvals.ts.

Line accounting: 4,430 after vs 4,407 before — the delta is import headers.
